requirements.txt
Flask==2.3.3
pymongo==4.6.0
python-dotenv==1.0.0
Flask-PyMongo==2.3.0
flask-bcrypt==1.0.1
flask-login==0.6.2
flask-wtf==1.2.1
email-validator==2.1.0.post1